challenges/jwt-forgery/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
PyJWT==2.8.0
python-multipart==0.0.6